+++ Tools.cu
@@ -8,10 +8,10 @@ int*   h_no_hits;
 int*   h_sensor_Zs;
 int*   h_sensor_hitStarts;
 int*   h_sensor_hitNums;
-int*   h_hit_IDs;
+unsigned int* h_hit_IDs;
 float* h_hit_Xs;
 float* h_hit_Ys;
-int*   h_hit_Zs;
+float* h_hit_Zs;
 
 void setHPointersFromInput(uint8_t * input, size_t size){
   uint8_t * end = input + size;
@@ -21,10 +21,10 @@ void setHPointersFromInput(uint8_t * input, size_t size){
   h_sensor_Zs        = (int32_t*)input; input += sizeof(int32_t) * *h_no_sensors;
   h_sensor_hitStarts = (int32_t*)input; input += sizeof(int32_t) * *h_no_sensors;
   h_sensor_hitNums   = (int32_t*)input; input += sizeof(int32_t) * *h_no_sensors;
-  h_hit_IDs          = (int32_t*)input; input += sizeof(int32_t) * *h_no_hits;
+  h_hit_IDs          = (uint32_t*)input; input += sizeof(uint32_t) * *h_no_hits;
   h_hit_Xs           = (float*)  input; input += sizeof(float)   * *h_no_hits;
   h_hit_Ys           = (float*)  input; input += sizeof(float)   * *h_no_hits;
-  h_hit_Zs           = (int32_t*)input; input += sizeof(int32_t) * *h_no_hits;
+  h_hit_Zs           = (float*)  input; input += sizeof(float)   * *h_no_hits;
 
   if (input != end)
     throw std::runtime_error("failed to deserialize event");
